% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIMLN-package.R
\docType{package}
\name{ZIMLN-package}
\alias{ZIMLN}
\alias{ZIMLN-package}
\title{ZIMLN: Zero-Inflated Multivariate Rounded Log-Normal Model for Microbiome Count Data}
\description{
A Bayesian zero-inflated multivariate rounded log-normal (ZI-MLN) model for next-generation sequencing microbiome abundance data. The model infers the interaction (covariance) structure between microbial features directly from observed counts in the presence of covariates and excess zeros, using a Dirichlet-Laplace joint-sparsity prior on a low-rank plus diagonal covariance decomposition and a probit zero-inflation component. Methods are described in Zhang, Shuler, Lee and Jeganathan (2023) \doi{10.1214/22-AOAS1690}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zsj950708/ZI-MLN}
  \item Report bugs at \url{https://github.com/Zsj950708/ZI-MLN/issues}
}

}
\author{
\strong{Maintainer}: Shuangjie Zhang \email{szhan209@tamu.edu}

}
\keyword{internal}
